implement movements

+++ TigersDen/src/main/java/TigersDen/BL/ConfigurationService/BussinessLogic/JsonConfigurationService.java
@@ -75,7 +75,7 @@ public class JsonConfigurationService implements IConfigurationService {
     }
 
     @Override
-    public float GetPieceMovementSpeed() {
+    public int GetPieceMovementSpeed() {
         return configurationData.getPieceMovementSpeed();
     }
 }
